Migrate to electron
While tauri was a nice idea, because of easy rust integration, it
turned out to have drawbacks as well.

Mainly performance on Linux is abysimal and I am developing on linux.
Also the runtime isn't always the same, so more testing would be needed
to ensure everything is working correctly in all webviews.